Move cbor. Fix acceptance tests
